requirements.txt
flask
groq
python-dotenv
gtts
pydub
gunicorn
